try another way to use generated client

+++ packages/generator-prisma/src/generator.ts
@@ -1,7 +1,9 @@
 import { GeneratorConfig } from "@prisma/generator-helper";
+import * as esbuild from 'esbuild';
 import { glob } from "glob";
 import fs from "node:fs";
 import path from "node:path";
+import * as ts from "typescript";
 import { getEnvValue } from "./env";
 
 export const getNewPrismaClientGenerator = (generators: GeneratorConfig[]) => {
@@ -11,9 +13,43 @@ export const getNewPrismaClientGenerator = (generators: GeneratorConfig[]) => {
   });
 };
 
-export const updateNextAdminPrismaTypesImport = (
+const buildGeneratedFilesTo = async (filesToBuild: string[], outDir: string) => {
+  await Promise.all((['cjs', 'esm'] as esbuild.Format[]).map(async (format) => {
+    const ext = format === 'cjs' ? '.js' : '.mjs';
+
+    return esbuild.build(
+      {
+        entryPoints: filesToBuild,
+        bundle: false,
+        outdir: outDir,
+        format,
+        outExtension: {
+          '.js': ext
+        }
+      },
+    )
+  }))
+}
+
+const buildGeneratedDtsFiles = async (files: string[], outDir: string) => {
+  const options: ts.CompilerOptions = {
+    allowJs: true,
+    declaration: true,
+    emitDeclarationOnly: true,
+    outDir
+  }
+  const host = ts.createCompilerHost(options);
+
+  host.writeFile = (fileName, contents) => {
+    fs.writeFileSync(fileName, contents);
+  }
+
+  const program = ts.createProgram(files, options, host);
+  program.emit();
+}
+
+export const updateNextAdminPrismaTypesImport = async (
   generator: GeneratorConfig,
-  pathReplacement?: string
 ) => {
   const nextAdminPath = path.dirname(
     require.resolve("@premieroctet/next-admin")
@@ -21,12 +57,38 @@ export const updateNextAdminPrismaTypesImport = (
 
   const generatorOutput = getEnvValue(generator.output!) ?? "";
 
+  if (!fs.existsSync(path.resolve(generatorOutput))) {
+    throw new Error(`Generator output path ${generatorOutput} does not exist, make sure to declare your prisma client BEFORE Next-Admin's one`);
+  }
+
+  const generatedPrismaInNextAdmin = path.join(nextAdminPath, 'generated', 'prisma');
+
+  const generatedFiles = glob.sync("**/*.{js,mjs,ts,mts}", { cwd: generatorOutput, absolute: true })
+
+  await buildGeneratedFilesTo(
+    generatedFiles,
+    generatedPrismaInNextAdmin
+  )
+
+  await buildGeneratedDtsFiles(
+    generatedFiles,
+    generatedPrismaInNextAdmin,
+  )
+
   glob.sync("**/*.{js,mjs,ts}", { cwd: nextAdminPath }).forEach((file) => {
     const filePath = path.join(nextAdminPath, file);
     const fileContent = fs.readFileSync(filePath, "utf8");
+    const isEsm = file.endsWith(".mjs");
+
+    let pathReplacement = path.relative(path.dirname(filePath), generatedPrismaInNextAdmin);
+
+    if (!pathReplacement.startsWith(`.${path.sep}`)) {
+      pathReplacement = `.${path.sep}${pathReplacement}`;
+    }
+
     const updatedContent = fileContent.replace(
       /('|")@prisma\/client('|")/g,
-      `"${pathReplacement ?? path.relative(path.dirname(filePath), generatorOutput)}/client"`
+      `"${pathReplacement}/client${isEsm ? ".mjs" : ""}"`
     );
     fs.writeFileSync(filePath, updatedContent, "utf8");
   });
